Add util for creating extendable ESLint configurations
