+++ packages/base/src/index.test.js
@@ -1,7 +1,7 @@
 import { ESLint } from 'eslint';
 import { describe, it, expect } from 'vitest';
 
-import config from '.';
+import config from './index.js';
 
 describe('index', () => {
   it('is a valid ESLint config', async () => {
