+++ packages/typescript/src/index.test.mjs
@@ -3,7 +3,7 @@ import globals from 'globals';
 import { resolve } from 'path';
 import { describe, it, expect } from 'vitest';
 
-import config from '.';
+import config from './index.mjs';
 
 describe('index', () => {
   it('is a valid ESLint config', async () => {
